@@ -1,18 +1,23 @@
 ﻿using BirthdayPokemonCore.Data.Enums;
+using BirthdayPokemonCore.Models;
 using BirthdayPokemonCore.Repo;
 using BirthdayPokemonCore.Services;
 using BirthdayPokemonTestXunit.Tests;
 
 namespace BirthdayPokemonTestXunit
 {
-    public class PokemonServiceTests
+    public class FakePokemonServiceTests
     {
         private readonly PokemonService _service;
 
-        public PokemonServiceTests()
+        public FakePokemonServiceTests()
         {
+            // use the fake repo for testing
             var mockRepo = new FakePokeAPIRepo();
-            _service = new PokemonService(mockRepo);
+
+            //and use in-memory logger
+            InMemoryLogger<PokemonService> logger = new();
+            _service = new PokemonService(mockRepo, logger);
         }
 
         [Theory]
@@ -29,8 +34,9 @@ namespace BirthdayPokemonTestXunit
             int normalized = _service.CalculateNormalizedBirthdayNumber(birthday, EPokemonBirthdayCalculationType.Standard);
             Assert.Equal(expectedDex, normalized);
 
-            string name = await _service.GetPokemonNameByDexAsync(normalized);
-            Assert.Equal(expectedName, name);
+            PokemonInfo? pokemon = await _service.GetPokemonInfoByDexAsync(normalized);
+            Assert.NotNull(pokemon);
+            Assert.Equal(expectedName,  pokemon.Name);
         }
 
         [Theory]
@@ -53,28 +59,6 @@ namespace BirthdayPokemonTestXunit
         public void InvalidDates_ThrowOrFailGracefully(string date)
         {
             Assert.ThrowsAny<Exception>(() => DateTime.Parse(date));
-        }
-
-        [Fact(Skip = "Requires internet connection; live API call.")]
-        public async Task BirthdayPokemon_LiveApi_31_03_2025_Works()
-        {
-            var repo = new PokeAPIRepo();
-            var service = new PokemonService(repo);
-
-            //inputs
-            var format = EFormatType.DayMonth;
-
-            // tests
-            DateOnly birthday = new(2025, 03, 31);
-            var pokemon = await service.GetBirthdayPokemonInfoAsync(birthday, format, EPokemonBirthdayCalculationType.Standard);
-
-            Assert.NotNull(pokemon);
-
-            // results
-            Assert.InRange(pokemon.Info!.DexNumber, 1, 1025);
-            Assert.False(string.IsNullOrWhiteSpace(pokemon.Info.Name));
-
-            Console.WriteLine($"[LIVE TEST] 31/03/2025 -> #{pokemon.Info.DexNumber}: {pokemon.Info.Name}");
         }
 
         [Fact]
