@@ -2,6 +2,7 @@
 using BirthdayPokemonCore.Data.Extensions;
 using BirthdayPokemonCore.Interfaces;
 using BirthdayPokemonCore.Models;
+using Microsoft.Extensions.Logging;
 using System.Text;
 
 namespace BirthdayPokemonCore.Services
@@ -9,7 +10,7 @@ namespace BirthdayPokemonCore.Services
     /// <summary>
     /// Service to determine Pokemon Dex number based on birthday.
     /// </summary>
-    public class PokemonService(IPokemonRepository pokemonRepository) : IPokemonService
+    public class PokemonService(IPokemonRepository pokemonRepository, ILogger<PokemonService> logger) : IPokemonService
     {
         private const int MaxDexNumber = 1025;
 
@@ -19,7 +20,6 @@ namespace BirthdayPokemonCore.Services
         /// <param name="birthday">The birthday date.</param>
         /// <param name="formatused">The format used for the birthday number calculation.</param>
         /// <returns>An integer representing the birthday number in the specified format.</returns>
-        /// <exception cref="ArgumentException">Thrown when an invalid format is specified.</exception>
         /// <remarks>Defaults to MM/DD format if none specified.</remarks>
         public int GetBirthdayNumber(DateOnly birthday, EFormatType formatused = EFormatType.DayMonth)
         {
@@ -35,20 +35,22 @@ namespace BirthdayPokemonCore.Services
                 }
                 else
                 {
-                    throw new ArgumentException("Invalid format specified. Use 'DD/MM' or 'MM/DD'.");
+                    logger.LogError("Invalid format specified: {formatused}. Use 'DD/MM' or 'MM/DD'.", formatused);
+                    return -1;
                 }
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                throw;
+                logger.LogError(ex, "Error calculating birthday number for date {birthday} with format {formatused}.", birthday, formatused);
+                return -1;
             }
         }
 
 
-        public async Task<string> ExportBirthdayPokemonDistributionToCsvAsync(string filePath = "birthday_pokemon_stats.csv")
+        public async Task<string> ExportBirthdayPokemonDistributionToCsvAsync(string filePath = "birthday_pokemon_stats.csv", EPokemonBirthdayCalculationType calculationType = EPokemonBirthdayCalculationType.Standard)
         {
-            var stats = await GenerateBirthdayPokemonDistributionAsync();
-            var sb = new StringBuilder();
+            PokemonDictionaryDistributation stats = await GenerateBirthdayPokemonDistributionAsync(calculationType);
+            StringBuilder sb = new ();
 
             sb.AppendLine("DexNumber,Name,Count");
 
@@ -61,59 +63,105 @@ namespace BirthdayPokemonCore.Services
             return Path.GetFullPath(filePath);
         }
 
+        /// <summary>
+        /// <para>Only for console applications</para>
+        /// Explains the birthday Pokémon based on input date string and format.
+        /// </summary>
+        /// <param name="input">
+        /// The input date string representing the birthday.
+        /// </param>
+        /// <param name="formatUsed">
+        /// The format used for the birthday number calculation.
+        /// </param>
+        /// <param name="birthdayCalculationType">
+        /// The type of birthday calculation to use (Standard, Seed, YearBased).
+        /// </param>
+        /// <returns>
+        /// A string explanation of the birthday Pokémon calculation and result.
+        /// </returns>
         public async Task<string> ExplainBirthdayPokemonAsync(string input, EFormatType formatUsed, EPokemonBirthdayCalculationType birthdayCalculationType = EPokemonBirthdayCalculationType.Standard)
         {
             try
             {
                 if (string.IsNullOrWhiteSpace(input))
-                    throw new ArgumentException("Input date cannot be null or empty.");
+                {
+                    logger.LogError("Input date string is null or empty.");
+                    return string.Empty;
+                }
 
-                var explanation = new StringBuilder();
 
                 // Step 1: Parse and detect format
                 DateOnly birthday = input.ParseDateWithFormat(formatUsed);
-                explanation.AppendLine($"Parsed date: {birthday:dd/MM/yyyy}");
 
-                // Step 1.5: Use specified format
-                explanation.AppendLine($"Using specified format: {formatUsed.ToString()}");
 
                 // Step 3: Normalize to valid Pokédex range
-                var (normalized, output) = CalculateNormalizedBirthdayNumberWithExplanation(birthday, formatUsed, birthdayCalculationType);
-                explanation.AppendLine(output);
+                (int normalized, string output) = CalculateNormalizedBirthdayNumberWithExplanation(birthday, formatUsed, birthdayCalculationType);
+
 
                 // Step 4: Lookup Pokémon
-                string name = await GetPokemonNameByDexAsync(normalized);
-                explanation.AppendLine($"Result -> #{normalized}: {name}");
+                PokemonInfo? pokemon = await GetPokemonInfoByDexAsync(normalized);
 
-                return explanation.ToString();
+                if (pokemon == null)
+                {
+                    logger.LogError("No Pokémon found for normalized Dex number: {normalized}.", normalized);
+                    return string.Empty;
+                }
+
+                if(string.IsNullOrEmpty(pokemon.Name))
+                {
+                    logger.LogError("Pokémon name is empty for normalized Dex number: {normalized}.", normalized);
+                    return string.Empty;
+                }
+
+                var colorStringBuilder = birthday.ToExplainationString(formatUsed, normalized, pokemon.Name);
+
+                string plaintextExplanation = colorStringBuilder.ToString();
+
+                if(string.IsNullOrEmpty(plaintextExplanation))
+                {
+                    logger.LogError("Generated explanation is empty for input: {input} with format: {formatUsed} and calculation type: {birthdayCalculationType}.", input, formatUsed, birthdayCalculationType);
+                    return string.Empty;
+                }
+
+                //only for console applications
+                bool is_console = Console.OpenStandardInput(1) != Stream.Null;
+                if (is_console)
+                {
+                    colorStringBuilder.WriteToConsole();
+                }
+                return plaintextExplanation;
             }
             catch (Exception ex)
             {
-                return $"Error explaining birthday Pokémon: {ex.Message}";
+                //log the exception
+                logger.LogError(ex, "Error explaining birthday Pokémon for input: {input} with format: {formatUsed} and calculation type: {birthdayCalculationType}.", input, formatUsed, birthdayCalculationType);
+                return string.Empty;
             }
         }
 
-        public int CalculateNormalizedBirthdayNumber(DateOnly birthday, EPokemonBirthdayCalculationType calculationType)
+        public int CalculateNormalizedBirthdayNumber(DateOnly birthday, EPokemonBirthdayCalculationType calculationType, EFormatType formatType = EFormatType.DayMonth)
         {
             if (calculationType == EPokemonBirthdayCalculationType.Standard)
             {
-                int raw = GetBirthdayNumber(birthday);
+                int raw = GetBirthdayNumber(birthday, formatType);
 
                 return NormalizeDexNumber(raw);
             }
             else if (calculationType == EPokemonBirthdayCalculationType.Seed)
             {
                 // Seed calculation logic here
-                return CalculateNumberWithBirthday(birthday);
+                (int result, string explaination) result = CalculateNumberWithBirthdayWithSteps(birthday);
+                return result.result;
             }
             else if (calculationType == EPokemonBirthdayCalculationType.YearBased)
             {
                 // Year-based calculation logic here
-                return CalculateNumberWithYear(birthday);
+                return CalculateNumberWithYear(birthday, formatType);
             }
             else
             {
-                throw new ArgumentException("Invalid calculation type specified.");
+                logger.LogError("Invalid calculation type specified: {calculationType}.", calculationType);
+                return -1;
             }
         }
 
@@ -133,8 +181,6 @@ namespace BirthdayPokemonCore.Services
                 {
                     output.AppendLine($"Within Pokédex range: {normalized}");
                 }
-
-
                 return (normalized, output.ToString());
             }
             else if (calculationType == EPokemonBirthdayCalculationType.Seed)
@@ -146,13 +192,22 @@ namespace BirthdayPokemonCore.Services
             else if (calculationType == EPokemonBirthdayCalculationType.YearBased)
             {
                 // Year-based calculation logic here
-                int result = CalculateNumberWithYear(birthday);
-                string output = $"Year based calculation (day * month % year) = ({birthday.Day} * {birthday.Month} % {birthday.Year}). Resulted: {result}";
-                return (result, output);
+                int result = CalculateNumberWithYear(birthday, formatUsed);
+                if(formatUsed == EFormatType.DayMonth)
+                {
+                    string output = $"Year based calculation (day * month % year) = ({birthday.Day} * {birthday.Month} % {birthday.Year}). Resulted: {result}";
+                    return (result, output);
+                }
+                else
+                {
+                    string output = $"Year based calculation (month * day % year) = ({birthday.Month} * {birthday.Day} % {birthday.Year}). Resulted: {result}";
+                    return (result, output);
+                }
             }
             else
             {
-                throw new ArgumentException("Invalid calculation type specified.");
+                logger.LogError("Invalid calculation type specified: {calculationType}.", calculationType);
+                return (-1, string.Empty);
             }
         }
 
@@ -162,202 +217,216 @@ namespace BirthdayPokemonCore.Services
             try
             {
                 if (number <= 0)
+                {
                     throw new ArgumentOutOfRangeException(nameof(number), "Dex number cannot be 0 or negative.");
+                }
 
                 if (number <= MaxDexNumber)
+                {
                     return number;
+                }
 
                 // Proper modular wrap for any range
                 return (number - 1) % MaxDexNumber + 1;
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                throw;
+                logger.LogError(ex, "Error normalizing Dex number: {number}.", number);
+                return -1;
             }
-        }
-
-        public static int CalculateNumberWithBirthday(DateOnly birthday)
-        {
-            var date = birthday;
-            int year = date.Year;
-            int month = date.Month;
-            int day = date.Day;
-
-            int baseYear = year % 2 == 0 ? year - 2 : year - 1;
-            DateOnly baseDate = new(baseYear, month, day);
-            var yearDigits = year.ToString().Select(c => int.Parse(c.ToString())).ToArray();
-            var seed = yearDigits[3] * 100 + yearDigits[2];
-            var datediff = date.DayOfYear - baseDate.DayOfYear;
-            var result = (datediff / 86400000 + seed) % MaxDexNumber + 1;
-            return result;
         }
 
         public static (int result, string explaination) CalculateNumberWithBirthdayWithSteps(DateOnly birthday)
         {
-            var date = birthday;
+            DateOnly date = birthday;
             int year = date.Year;
             int month = date.Month;
             int day = date.Day;
 
             int baseYear = year % 2 == 0 ? year - 2 : year - 1;
             DateOnly baseDate = new(baseYear, month, day);
-            var yearDigits = year.ToString().Select(c => int.Parse(c.ToString())).ToArray();
-            var seed = yearDigits[3] * 100 + yearDigits[2];
-            var datediff = date.DayOfYear - baseDate.DayOfYear;
-            var result = (datediff / 86400000 + seed) % MaxDexNumber + 1;
+            int[] yearDigits = year.ToString().Select(c => int.Parse(c.ToString())).ToArray();
+            int seed = yearDigits[3] * 100 + yearDigits[2];
+            int datediff = date.DayOfYear - baseDate.DayOfYear;
+            int result = (datediff / 86400000 + seed) % MaxDexNumber + 1;
 
             //write explanation log with whole calculation steps
-            StringBuilder explanation = new();
-            //How we calculated baseyear
-            explanation.AppendLine($"Base Year Calculation: {(year % 2 == 0 ? $"{year} is even, so base year is {year - 2}" : $"{year} is odd, so base year is {year - 1}")}");
-
-            //how we calculated seed
-            explanation.AppendLine($"Seed Calculation: Last two digits of year {year} are {yearDigits[2]} and {yearDigits[3]}, so seed is ({yearDigits[3]} * 100) + {yearDigits[2]} = {seed}");
-
-            explanation.AppendLine($"Date Difference Calculation: Day of year for {date:dd/MM/yyyy} is {date.DayOfYear}, Day of year for base date {baseDate:dd/MM/yyyy} is {baseDate.DayOfYear}, so difference is {date.DayOfYear} - {baseDate.DayOfYear} = {datediff}");
-
-            explanation.AppendLine($"Final Calculation: (((( {datediff} ) / 86400000) + {seed}) % {MaxDexNumber}) + 1 = {result}");
-
-            return (result, explanation.ToString());
+            string explanation = birthday.ToExplainationString(MaxDexNumber, result);
+            return (result, explanation);
         }
 
-        public static int CalculateNumberWithYear(DateOnly birthday)
+        public int CalculateNumberWithYear(DateOnly birthday, EFormatType formatType = EFormatType.DayMonth)
         {
-            int day = birthday.Day;
-            int month = birthday.Month;
-            int year = birthday.Year;
-
             try
             {
-                DateValidation(day, month, year);
+                birthday.DateValidation();
             }
             catch (Exception ex)
             {
-                throw new ArgumentException($"Date validation failed: {ex.Message}");
+                logger.LogError(ex, "Invalid date provided for year-based calculation: {birthday}.", birthday);
+                return -1;
             }
 
-            return day * month % year;
-        }
-
-        public async Task<string> GetPokemonNameByDexAsync(int dexNumber)
-        {
-            try
+            return formatType switch
             {
-                var pokemon = await pokemonRepository.GetPokemonNameAsync(dexNumber);
-                if (string.IsNullOrEmpty(pokemon))
-                {
-                    return $"Pokémon #{dexNumber}";
-                }
-
-                return pokemon;
-            }
-            catch
-            {
-                return $"Pokémon #{dexNumber}";
-            }
+                EFormatType.DayMonth => birthday.Day * birthday.Month % birthday.Year,
+                _ => birthday.Month * birthday.Day % birthday.Year
+            };
         }
 
         public async Task<PokemonInfo?> GetPokemonInfoByDexAsync(int dexNumber)
         {
             try
             {
-                var pokemon = await pokemonRepository.GetPokemonInfoAsync(dexNumber);
+                PokemonInfo pokemon = await pokemonRepository.GetPokemonInfoAsync(dexNumber);
+                if(string.IsNullOrEmpty(pokemon.Name))
+                {
+                    logger.LogWarning("Pokémon name is empty for Dex number: {dexNumber}.", dexNumber);
+                    return null;
+                }
+
+                if(string.IsNullOrEmpty(pokemon.ImageUrl))
+                {
+                    logger.LogWarning("Pokémon image URL is empty for Dex number: {dexNumber}.", dexNumber);
+                }
+
+                if(pokemon.DexNumber <= 0)
+                {
+                    logger.LogWarning("Pokémon Dex number is invalid ({dexNumber}) for Dex number: {dexNumber}.", pokemon.DexNumber, dexNumber);
+                    return null;
+                }
                 return pokemon;
             }
-            catch
+            catch(Exception ex)
             {
-                return new PokemonInfo($"Pokémon #{dexNumber}", string.Empty, dexNumber);
+                logger.LogError(ex, "Error fetching Pokémon info for Dex number: {dexNumber}.", dexNumber);
+                return null;
             }
         }
 
-        public async Task<PokemonInfoWithSteps> GetBirthdayPokemonInfoAsync(DateOnly birthday, EFormatType formatType, EPokemonBirthdayCalculationType calculationType = EPokemonBirthdayCalculationType.Standard)
-        {
-            (int normalizedDex, string output) = CalculateNormalizedBirthdayNumberWithExplanation(birthday, formatType, calculationType);
-            var pokemonInfo = await GetPokemonInfoByDexAsync(normalizedDex);
-            string explanation = GeneratePokemonInfoExplanation(birthday, formatType, normalizedDex, pokemonInfo, output, calculationType);
-            return new PokemonInfoWithSteps(pokemonInfo, explanation);
-        }
-
-        private static bool DateValidation(int day, int month, int year)
+        public async Task<PokemonInfoWithSteps?> GetBirthdayPokemonInfoAsync(DateOnly birthday, EFormatType formatType, EPokemonBirthdayCalculationType calculationType = EPokemonBirthdayCalculationType.Standard)
         {
             try
             {
-                //validate every edge cases with day, month and year
-                if (day == 0)
+                (int normalizedDex, string output) = CalculateNormalizedBirthdayNumberWithExplanation(birthday, formatType, calculationType);
+                if (normalizedDex <= 0)
                 {
-                    throw new ArgumentException("Day cannot be zero.");
+                    logger.LogError("Calculated normalized Dex number is invalid: {normalizedDex} for birthday: {birthday}, format: {formatType}, calculation type: {calculationType}.", normalizedDex, birthday, formatType, calculationType);
+                    return null;
                 }
 
-                if (month == 0)
+                if (string.IsNullOrEmpty(output))
                 {
-                    throw new ArgumentException("Month cannot be zero.");
+                    logger.LogError("Calculation output is empty for birthday: {birthday}, format: {formatType}, calculation type: {calculationType}.", birthday, formatType, calculationType);
+                    return null;
                 }
 
-                if (year == 0)
-                {
-                    throw new ArgumentException("Year cannot be zero.");
+                PokemonInfo? pokemonInfo = await GetPokemonInfoByDexAsync(normalizedDex);
+                if(pokemonInfo == null) {
+                    logger.LogError("No Pokémon info found for normalized Dex number: {normalizedDex}.", normalizedDex);
+                    return null;
                 }
-
-                if (day > 31)
-                {
-                    throw new ArgumentException("Day cannot be greater than 31.");
-                }
-
-                if (month > 12)
-                {
-                    throw new ArgumentException("Month cannot be greater than 12.");
-                }
-
-                if (year < 1850 && year > 2100)
-                {
-                    throw new ArgumentException("Year must be between 1850 and 2100.");
-                }
-                return true;
+                string explanation = pokemonInfo.ToExplaniationString(birthday, formatType, calculationType, output, normalizedDex);
+                return new PokemonInfoWithSteps(pokemonInfo, explanation);
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                throw;
+                logger.LogError(ex, "Error getting birthday Pokémon info for birthday: {birthday}, format: {formatType}, calculation type: {calculationType}.", birthday, formatType, calculationType);
+                return null;
             }
         }
 
-
-        private static string GeneratePokemonInfoExplanation(DateOnly birthday, EFormatType formatType, int normalizedDex, PokemonInfo? pokemonInfo, string outputfromcalculation, EPokemonBirthdayCalculationType calculationType = EPokemonBirthdayCalculationType.Standard)
+        public async Task<PokemonDictionaryDistributation> GenerateBirthdayPokemonDistributionAsync(EPokemonBirthdayCalculationType calculationType = EPokemonBirthdayCalculationType.Standard)
         {
-            var explanation = new StringBuilder();
-            explanation.AppendLine($"Parsed date: {birthday:dd/MM/yyyy}");
-            explanation.AppendLine($"Using specified format: {formatType.ToString()}");
-            explanation.AppendLine($"Using calculation type: {calculationType}");
-            explanation.AppendLine(outputfromcalculation);
-            explanation.AppendLine($"Result -> #{normalizedDex}: {pokemonInfo?.Name}");
-            return explanation.ToString();
-        }
+            Dictionary<int, PokemonDistributionEntry> results = [];
+            results.Clear();
 
-        public async Task<PokemonDictionaryDistributation> GenerateBirthdayPokemonDistributionAsync()
-        {
-            var results = new Dictionary<int, PokemonDistributionEntry>();
+            // Iterate through all days of the year (leap year)
+            var currentYear = DateTime.Now.Year;
+            logger.LogInformation("Generating birthday Pokémon distribution for year: {currentYear}", currentYear);
 
             for (int month = 1; month <= 12; month++)
             {
-                for (int day = 1; day <= DateTime.DaysInMonth(2024, month); day++)
+                // Get number of days in the month
+                for (int day = 1; day <= DateTime.DaysInMonth(currentYear, month); day++)
                 {
-                    DateOnly date = new(2024, month, day);
-                    int normalized = CalculateNormalizedBirthdayNumber(date, EPokemonBirthdayCalculationType.Standard);
+                    // Create DateOnly for the current day
+                    DateOnly date = new(currentYear, month, day);
 
-                    if (!results.TryGetValue(normalized, out PokemonDistributionEntry? value))
+                    logger.LogInformation("Processing date: {date:dd/MM/yyyy}", date);
+
+                    // Calculate normalized Dex number
+                    int normalizedForDayMonth = CalculateNormalizedBirthdayNumber(date, calculationType);
+                    int normalizedForMonthDay = CalculateNormalizedBirthdayNumber(date, calculationType, EFormatType.MonthDay);
+
+                    // Process Day/Month format
+                    if (!results.TryGetValue(normalizedForDayMonth, out PokemonDistributionEntry? value))
                     {
-                        string name = await GetPokemonNameByDexAsync(normalized);
+                        // Fetch Pokémon info
+                        PokemonInfo? pokemon = await GetPokemonInfoByDexAsync(normalizedForDayMonth);
+                        if (pokemon == null)
+                        {
+                            logger.LogWarning("No Pokémon found for normalized Dex number: {normalized} during distribution generation.", normalizedForDayMonth);
+                            continue;
+                        }
+
+                        if(string.IsNullOrEmpty(pokemon.Name))
+                        {
+                            logger.LogWarning("Pokémon name is empty for normalized Dex number: {normalized} during distribution generation.", normalizedForDayMonth);
+                            continue;
+                        }
+
+                        // Initialize new entry
                         value = new PokemonDistributionEntry
                         {
-                            DexNumber = normalized,
-                            Name = name,
+                            DexNumber = normalizedForDayMonth,
+                            Name = pokemon.Name,
                             Count = 0
                         };
-                        results[normalized] = value;
+                        // Add to results
+                        results[normalizedForDayMonth] = value;
+                        logger.LogInformation("Added new Pokémon to distribution: {Name} (Dex #{DexNumber})", value.Name, value.DexNumber);
                     }
+
+                    // Increment count for this Pokémon
+                    logger.LogInformation("Incrementing count for Pokémon: {Name} (Dex #{DexNumber})", value.Name, value.DexNumber);
                     value.Count++;
+
+                    // Process Month/Day format
+                    if (!results.TryGetValue(normalizedForMonthDay, out PokemonDistributionEntry? valueMD))
+                    {
+                        // Fetch Pokémon info
+                        PokemonInfo? pokemonMD = await GetPokemonInfoByDexAsync(normalizedForMonthDay);
+                        if (pokemonMD == null)
+                        {
+                            logger.LogWarning("No Pokémon found for normalized Dex number: {normalized} during distribution generation.", normalizedForMonthDay);
+                            continue;
+                        }
+                        if (string.IsNullOrEmpty(pokemonMD.Name))
+                        {
+                            logger.LogWarning("Pokémon name is empty for normalized Dex number: {normalized} during distribution generation.", normalizedForMonthDay);
+                            continue;
+                        }
+                        // Initialize new entry
+                        valueMD = new PokemonDistributionEntry
+                        {
+                            DexNumber = normalizedForMonthDay,
+                            Name = pokemonMD.Name,
+                            Count = 0
+                        };
+                        // Add to results
+                        results[normalizedForMonthDay] = valueMD;
+                        logger.LogInformation("Added new Pokémon to distribution: {Name} (Dex #{DexNumber})", valueMD.Name, valueMD.DexNumber);
+                    }
+
+                    // Increment count for this Pokémon
+                    logger.LogInformation("Incrementing count for Pokémon: {Name} (Dex #{DexNumber})", valueMD.Name, valueMD.DexNumber);
+                    valueMD.Count++;
+
                 }
             }
 
+            logger.LogInformation("Completed generating birthday Pokémon distribution.");
             return new PokemonDictionaryDistributation
             {
                 Dictionary = results
