@@ -7,6 +7,6 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using BirthdayPokemonWeb
-@using BirthdayPokemonWeb.Layout
-@using BlazorBootstrap;
-@using BirthdayPokemonCore.Data.Enums
+@using BirthdayPokemonRazorCore
+@using BirthdayPokemonRazorCore.Layout
+@using BirthdayPokemonRazorCore.Pages
